Make autoloading with Zeitwerk work
