+++ lib/phlex.rb
@@ -1,22 +1,10 @@
 # frozen_string_literal: true
 
 require "active_support/core_ext/string/output_safety"
+require "zeitwerk"
 
-require_relative "phlex/version"
-require_relative "phlex/callable"
-require_relative "phlex/block"
-require_relative "phlex/tag"
-require_relative "phlex/tag/class_collector"
-require_relative "phlex/tag/void_element"
-require_relative "phlex/node"
-require_relative "phlex/tag/standard_element"
-require_relative "phlex/tags"
-require_relative "phlex/context"
-require_relative "phlex/cache_key"
-require_relative "phlex/cacheable"
-require_relative "phlex/cacheable_object"
-require_relative "phlex/component"
-require_relative "phlex/text"
+loader = Zeitwerk::Loader.for_gem
+loader.setup
 
 module Phlex
   extend self
@@ -49,4 +37,3 @@ module Phlex
     end
   end
 end
-
